Quit UI Check mode when composable renamed or deleted
When in UI Check mode, we react to changes to the original composable
from which the mode was started. But if the name of the composable is
changed, or if it cannot be previewed anymore (e.g. no more @Preview),
then it does not make sense to stay in UI Check mode.
Instead we go back to normal mode.

Bug: 304204594
Bug: 305909266
Test: checked manually
Change-Id: I2bbe6e97bf07b5d94a186c208b611858d3746320

GitOrigin-RevId: ede71d9b050b9ca0fb054b08e10dc5876f6f35aa