Handle OOB Speed Events
An OOB speed event is a non-zero speed event that occurs outside a
tracked connection. These can be caused non-HTTP network traffic or HTTP
traffic from an unsupported transport layer. These events cause spikes
in the timeline which do not have corresponding connection entries in
the ConnectionView. We should still probably display these in the
timeline, but we don't want to automatically extend the timeline when
they arrive.

The heuristics for updating the timeline due to a speed event is as
follows: Once a connection is started, all speed events extend the
timeline until the first zero-event with a timestamp later than (greater
than) the connection end time.

Bug: NA
Test: Added
Change-Id: I1a5b8d022eea0ed739b861b556fba95bff1cead6

GitOrigin-RevId: 3ba71c566a75bd49e72f546509a31e6fad4ed6b4