Retain last image even for failed renders
Adds a new COMPOSE_PREVIEW_KEEP_IMAGE_ON_ERROR flag that, when enabled,
allows the Compose renders to be kept even when a new render fails.
This allows for the user to see the last image even when there are fatal
errors that do not allow any rendering.

When the flag is enabled, the conditions that show the error panel also
change. If there is valid image, the panel is never shown. This has a
side effect in cases like using device rendering in which usually an
image is always returned although it might be blank when an error has
happend.

Test: LayoutlibSceneManagerTest
Bug: N/A
Change-Id: I210bbaa6135005dd2d334e7d3cbf8ec0ddb4caff

GitOrigin-RevId: a5f9c19e204caf4db70193eb3462e8a8400b1b32