Set correct data context for preview toolbar
Storing objects such as the preview manager or instance in the toolbar
actions can lead to leaks. To avoid this, such data should be retrieved
by the action through data context. To achieve this, the target
component of the toolbar must be the scene view, and not the surface, so
that it can correspond to one preview instance.
As the target component must be a Swing component, we use
SceneViewPeerPanel which we make into a DataProvider so that the correct
data can be accessed from the toolbar actions.
This allows for removing passing the data directly to the actions in
their constructor.
In addition, this refactors the way the toolbar is created to avoid code
duplication and use this to align the Layout Validation toolbar with the
one from Compose preview.

Bug: N/A
Test: existing tests
Change-Id: I5b58cb1bdeb41b726815ad8d9215f937d7382a80

GitOrigin-RevId: 75726d948e02958e22254dea20376356927faee0