Migrate to ApplicationInitializedListener#execute
In IntelliJ 2023.3, ApplicationInitializedListener.execute() logs an
error if not overridden, thereby breaking most tests.

Also: since ApplicationInitializedListener.execute() is a suspend
function, it cannot be implemented in Java. This is inconvenient for
certain modules such as aswb.jcgdi.blaze.plugin.aswb which are
Java-only. To accommodate this, JetBrains added a compatibility
class called ApplicationInitializedListenerJavaShim, which we
backport temporarily here.

Change-Id: I67017c10b657a9c09bba0741453da8709047d98e

GitOrigin-RevId: e5d005a0106dde341edd96c4dddcc10b5797fe96