Adjust UI model fields to give most updated value
Some of the UI models had getter methods for fields in sub-models
for convenience. Mistakenly, these were defined as normal fields
and thus only held the value they read at model instantiation time.
Thus, when the instance variable is called later, it gives an old
value. Thus, the fields get() method have been overriden to give
the most updated value.

In this change you will find that the "getSelectedTaskType" getter
method was removed as it was conflicitng with the existing implicit
getter given by kotlin.

Bug: 277797248
Bug: 304594618
Test: added
Change-Id: I8dcad884edb77ae91b88a1645dddd5dbd9c5c008

GitOrigin-RevId: c20d00cbf6fe5c3c093c5be014c59e3dc6ecf366