Refactor ConnectionsView For Easier Column Handling
In preparation for adding more columns, this CL makes it safer to add
columns to ConnectionsView.

The Refactor Usage Tracking

The current implementation has 2 places where it needs to know things
about the column: Its default visibility and its renderer.

This means that if we add columns to the enum, we would have to update
these 2 code locations.

Instead of this, make the column provide these values itself.

Some classes were extracted from ConnectionsView to make the code
cleaner and a new package was introduced to host this component.

Bug: NA
Test: Pure refactor
Change-Id: I12e0d346b713a1da7adf79147ecefc95bc59a10c

GitOrigin-RevId: 0b6d5dc1c3c6e97cf5ab2edf878c85f28a82fc3c