Handle resource updates with commented lines
ResourceFolderRepository's logic to handle PSI updates assumes that
changes to XML comments can't affect resource ids. This is not strictly
true, as is the case when the user selects an entire XML tag and
comments it out with the "block comment" action. In that case the PSI
update identifies a comment as being changed; but in fact it's a tag
being removed and replaced by a comment. The reverse operation has a
similar error.

This change detects such modification and schedules a rescan of the
resource file instead of ignoring the change.

Fixes: 302262716
Test: ResourceFolderRepositoryTest.commentAndUncommentTag
Change-Id: Id86dd780fbe6a89d6e87917e58c54a88e5aa6088

GitOrigin-RevId: b6fe1d681ba6dfda17935b60d1344cd77d0224c9