Do not recreate the actions in ComposeViewControlAction
Some of the actions in ComposeViewControlAction need access to the
surface or the compose preview manager. They can get that from the data
context, but we were not passing the context from the action button to
the drop down actions in DropDownAction.
If we fix that, the actions in drop downs, like the ones in
ComposeViewControlAction, can get all the required information from the
data context of the action event. They do not need to be explicitly
passed a context. That means we do not need to recreate the actions
whenever the popup is displayed.

Bug: N/A
Test: existing tests
Change-Id: I599b2e4bf1e9a398c52643b494f81965b73a6880

GitOrigin-RevId: 563eb8c010eb8eac80345eb80253700e8418b52b