Uniquely name AQI workbenches
Workbenches uses names as a sort of handle for propagating state changes
and unfortunately AQI uses the same name for both Crashlytics and Vitals
workbenches. This resulted in some weird but subtle misbehavior, such as
minimizing the distribution side window in Crashlytics will also
minimize the distribution window in Vitals. I strongly suspect this is
the cause of the bug below.

Test: manually verified by QA.
Fixes: 308558239
Change-Id: Ica83d56ef77574f3d8f17c7656b9db75f183cded

GitOrigin-RevId: 5aecdc47b5e914c6cf2fdc3373516ec5c409843b