Fix JarManager cache for single file jars
When populating the initial cache, for JARs that contain only a single
file, we would accidentally generate an immutable map preventing
additional entries from being added to that file cache.
This is due to how toMap() works in Kotlin.
This change moves to use an explicit MutableMap to ensure that this can
not happen.

Fixes: 308140478
Test: JarManagerTest
Change-Id: I0ba6bac7158025c0b67dddcd141031f1ef78dee8

GitOrigin-RevId: 2315f533db9bfb3569c7491372168fcfefaba5aa