Prevent startup memory tracing caused by legacy cpu tracing
Because no timestamps are set for the start and stop trace status
events issued by the Legacy CPU command handler, the events of
are not sorted in the correct (chronological) order. Thus, it so
happens that when the most recent is retrieved, it is grabbing
the start status event instead of the stop status event, causing
StudioProfilers::startupMemoryProfilingStarted to return true.

Thus, timestamps were added.

Fixes: 309685267
Test: updated
Change-Id: I53a590cf2e202697ed068be69aa4d4eaa1eba526

GitOrigin-RevId: 4ecf9ab36323d74741277128d6861cb3d8513e7d