Add RecordingList component
The RecordingList component contains the import and export
actions to allow the user to amend and export the recording
list entries, as well as a table showing all the previously
captured and imported task recordings. This component will
serve as the left panel of the Past Recordings tab, enabling
the user to select a recording before selecting the respective
task to launch the recording with.

NOTE: Because the import button is written in compose, which
makes passing in a JComponent difficult to the error dialog
api used in the import action. Thus, an adjustment was made
to the intellij ui message handler to make the swing component
param nullable. It was verified that the same behavior was
exhibited with or without using null (it does not need the
JComponent to render the dialog popup).

Bug: 277798530
Test: added
Change-Id: I827c426c33d99dcbefd01849858f57725e7b4e3d

GitOrigin-RevId: ffdea925e13df1a4a485752bedfd23d2038086d2