Fix Preview being stuck between two modes
When changing the PreviewMode, we're also setting a "restore" mode that
we'll transition to when leaving the new mode. The problem is we might
make quick mode changes by clicking on the preview toolbar actions, e.g.
clicking on "Animation Preview" then on "Interactive Preview", causing
the preview to be stuck between Animation and Interactive.

We should limit the modes that can be used as "restore" to the ones it
makes sense doing so: Default and Gallery.

Test: N/A. Manually verified.
Fixes: 307939428
Change-Id: If97665c1be637c56c56b208863339fd977c54875

GitOrigin-RevId: f2580064df214f0bfa52be3be92a4fa6ca7ba6f4