Move Baseline Profiler generation action to top
This commit moves the baseline profiler generation action in the gutter
popup menu to the top of the menu, as per design guidelines.

The PsiElement changes were necessary to tie the BPG action to the
executor actions, since executor actions are tied to the IDENTIFIER
PsiElement, not the preceding CLASS PsiElement.

Fixes: 309133522
Test: (Kt|Java)BaselineProfileRunLineMarkerContributorTest
Change-Id: I64aaa6aac880789e92d7d00679d37aecb1c01361

GitOrigin-RevId: f7d2652f0b7198d12144f879d5bd7b17502c34c1