Refactor some messages in ComposeBundle.properties
Also, since we are no longer applying highlighting by default to
State Reads, we do not need to worry as much about race conditions
in highlighting. If it becomes an issue we can revisit, but for
now we want to stick with the annotator because it's going to do
more than highlight.

Bug: 309027083
Bug: 225218822
Test: Existing tests
Change-Id: I253f3ddbc8ee8184f4db76afa313cf19ece0b094

GitOrigin-RevId: c1d661f601c8228c404b6de5fa82508197df28c3