Restore Compose Layout Inspector on APK projects
Break up the large getAppInspectorJar() token function into smaller
pieces.  Provide fallback implementations of those pieces for the case
when there is no Project System token, and in each case make the
Gradle implementation call the fallback implementation (so ensuring
that the behaviour of the Layout Inspector in Gradle projects and
non-Gradle projects is the same).  The token is left in so as to allow
ASwB to customize the behaviour of any of the pieces.

Bug: 309462683
Test: existing
Change-Id: I07daf4aa2d5638efd956d9cfd22435776d8eaa95

GitOrigin-RevId: 2bf960873aa255a62dff163e00154910f663832c