Set instance filter before onEnter/onExit on mode change
When changing the preview mode, we should set the instance filter before
calling onEnter/onExit, otherwise we would trigger unnecessary refreshes
to previews that won't be present in the newly set mode.

This was specially problematic when transitioning from Default mode to
any other mode, because we were refreshing all the preview in that mode
before transitioning, instead of refreshing just one (except for UI
check).

Fixes: 309965217
Test: current tests should pass + manually verified.
Change-Id: Ie05b5d029bc1bfe2f525a37c6df383dfcd6d04b2

GitOrigin-RevId: 1c0dc3991a2fa01cdb70c9ee94e85275a569c6b2