Quick-fix to update minCompileSdk to resolve dependency incompatibility
If a module has a Aar depenency that requires a higher minCompileSdk,
build task that checks the dependency Aar metadata fails. This change
creates an issue checker that parses the error message and present
a quickfix that updates the minCompileSdk in the module build file.
This checker only supports the case when AGP(CheckAarMetadataTask) recommend a higher minCompileSdk version which cannot exceeds the version supported by current AGP.

Bug: 308175247
Test: UpdateCompileSdkProcessorTest,
      AarDependencyCompatibilityIssueCheckerTest,
      AarDependencyCompatibilityIssueCheckerIntegrationTest
Change-Id: I25496c686a057b375a098535b0faea4f5a3cb023

GitOrigin-RevId: 67330182cf9602af641cf49d408c59c8502407a6