Do not show no errors label at the start of UI Check mode
On starting UI Check mode, this shows an empty surface until it is
populated with the final result, either the previews with errors, or a
label saying no error was found.
This avoids showing the no error label immediately, and seeing previews
appear and disappear while the UI Check analysis is run.

Bug: 308444924
Test: Checked manually
Change-Id: Ic371f6a4d66134cad997edadfc7e340b96738514

GitOrigin-RevId: 2438a79c522236f4ed5b5d8465f7b489a3612438