[android] Rename module android.sdktools.sdk-common.gradle => android.sdktools.sdk-common.gradle.rt
The android.sdktools.sdk-common.gradle module is a runtime module (or should be treated as one),
and it should follow a specific naming convention for runtime modules, such as adding a .rt
(short for runtime) suffix to its name.
The .rt suffix indicates that the module should be included at runtime.

GitOrigin-RevId: 2e132f3fba379ec6b88c857fdf3e7cf5fb31ffc2