Restore resolving the running version in app inspector support
A change in mid-2021 inadvertently removed the behavior of the app
inspector tab launcher to resolve the inspector jar corresponding to
the running version of the library, instead always requesting the
minimum version.  This probably had no visible consequences over the
last two and a half years because this code path is used only for the
androidx.work background task inspector, which has had no incompatible
changes for at least that long.

Bug: none filed
Test: none apparently
Change-Id: I3c577831e747c97b765f92c5f4992a653a7cffed

GitOrigin-RevId: 831b70f26c51636b3ea8073ad53a1d5ca4ebf433