Force reordering of process list on device selection
When the preferred process is set, a reorder of the process list
is triggered. However, if no device is selected, no process list
is mapped to to sort. Thus, if the preferred process was set
before the device selection was made, when the device selection
was eventually made, the process list was not sorted.

To fix this case, we can force a reorder of the process list when
the device selection is made.

Bug: 277798461
Test: added
Change-Id: Ic30ff5d08f3a0bc947970c004e8f70a9ed7ecee5

GitOrigin-RevId: 642fd9ba985f7746b3199f1dbec6a6c731055455