Overhaul the quickfix machinery
This is the IDE-side of the switch over to a unified quickfix handler
for lint; see the commit message in the same topic for more.

In addition to implementing the IDE implementation of the lint fix
performer and quickfix, it removes various inspection quickfixes that
are now implemented on the lint side instead, fixes various bugs (for
example, DomPsiConverter's getElementsByTagName was modifying sibling
pointers, the getLastChild method was accidentally returning the first
child, fixes intention preview for a few inspections, and adds a bunch
of tests.

Bug: Not filed
Test: Unit tests included
Change-Id: Ic861cd3c241d7f5ce52550fee2d65e94bc09a79e

GitOrigin-RevId: fccf5ad6a1d1e7a85b01854a5fcf498fd7e92960