Add "No tasks available" if no supported tasks are available
In each past recording list entry, the last columb shows the
available tasks to open the recording with. If there are no
available tasks for a given recording, we were showing nothing
(empty string). However, no we will show "No tasks available"
so that the user knows why no tasks are enabled when they select
the entry. Ideally, we would also disable selection of such
entry, but with the table component being revamped in the future,
it might be a waste of time to do that now.

Also, it is worth acknowleding that currently we analyze the child
artifacts to see what tasks are supported, and thus a standalone
session item will show no supported tasks as of now. However, in
the future, the Live Task will only contains a stand alone session
item (no child artifact), therefore necessitatiing an adjustment
to the test added in this change.

Bug: 277798530
Test: added
Change-Id: I740692bc9f8537124fed5ddea6ce120654032a0d

GitOrigin-RevId: 95c3ec5076bb445b5866b1f9209a59859254852f