Prevent null PSI during validation
With the LE event queue, it is possible to encounter a race condition
between the retried compile for a file and a new event for that file.
The snapshot of PSI taken for the new event generated may be used by the
retried compile, causing the compilation of the new event to encounter
no snapshot.

To limit this, we move the enqueuing of the new event to
inside the write action of the change (preventing it from being added
during an active compilation) and change the exception thrown by the
null PSI to a warning. This only adds risk during the first change,
where if the race condition triggers an incompatible change may not be
flagged as such (changes to constructors/init/static init).

This can be fixed with a more comprehensive refactoring of the LE event
queue, but that is out of scope for this change.

Bug: 311041949
Test: Manual
Change-Id: Ibc25ca2efe959664b2040cdf253022c896203a75

GitOrigin-RevId: 91034b26be55680fe5a6c183043835a966f602ce