Clear selected tab when RD tool window is collapsed
This change adjust the behavior of embedded Layout Inspector
so that when the Running Device tool window is closed, the selected
tab is cleared and the Layout Inspector UI is removed from the RD
tab.
Layout Inspector UI is re-injected when the tool window is opened.

This removes any assumptions on what the RD tool window does with
its tabs when it collapsed.

Test: LayoutInspectorManagerTest, RunningDevicesStateObserverTest
Bug: 311167778
Change-Id: I42f92fde37f67ec643f2bde769ef7f281b2fa5c8

GitOrigin-RevId: c733829f7d2355d7c9e9eacbb0c19f313f170ee4