Fix layout management in Compose preview
Setting the layout manager when changing mode is done in the collect
method for the mode flow, it does not need to also be done in the
onEnter method.
In addition, sometimes it is done in the UI thread, and sometimes not.
This makes sure that the operations that should be done in the UI thread
are.
This also updates ComposePreviewRepresenationTest to wait for the layout
manager to be set instead of asserting that it was, as it can take some
time to propagate through the flow.

Bug: 311245856
Test: tests updated
Change-Id: Ied42df11785088102d0f6e3015cce00d3cd39f71

GitOrigin-RevId: 76ea47a025f2372b4e3870f5b89ff34642c30a15