Only call onAfterRender if actually rendering preview
After preview refresh, we call the onAfterRender callback. However, the
refresh method can be called if no previews are rendered, e.g. when
initializing the preview panel. Since the onAfterRender callback handles
logic that assume there are previews in the panel (e.g. so it indirectly
sets the appropriate zoom level when previews are rendered), we should
not trigger the callback if there are no previews in the panel.

Fixes: 311056411
Test: N/A. Visually verified that zoom-to-fit is applied in new files.
Change-Id: If33c29edf05d86e2484046a0c3043a627b4a1537

GitOrigin-RevId: 46fc9ae00a7c3440da8b5f3c4e4dffe210d4766c