Add task action bar and combine task based ux components
This change adds in a simple container component to wrap the
enter task button called the TaskActionBar. It is named as
such because there will be more components that will allow the
user to configure the task within it. However, the wrapping of
this component was an intermediate step in this change's larger
goal of combining all the pre-made task based ux components
into the tool window.

The aforementioned task action bar was combined to be placed
below the task grid into a new component: TaskGridAndActionBar.

The ProcessList component and TaskGridAndActionBar were combined
using a HorizontalSplitLayout (allows for a re-adjustable vertical
divider) to make the home tab's content. While RecordingList and
TaskGridAndActionBar components were combined in a similar fashion
to create the past recordings tab's content.

Note: the unnecessary ideProfilerComponents parameter has been removed
in this change. It was used for debugging purposes before, but is
no longer needed.

Bug: 277797248
Bug: 304594618
Fixes: 277797248
Test: added
Change-Id: Ic77606dbcf2630b29a40c4599e134ccb9639a624

GitOrigin-RevId: 07196fbb075bc375ab0e7ad73f0b1a4364dba92d