Add `logDirect` logs to resource traces
When resource update traces are dumped, they give a history of all the
resource-related trace lines that have been logged up until that point.
But some of the lines are logged using `logDirect`, which directly
outputs them to idea.log when that message is called and does not put
them in the buffer that may be dumped later. This makes debugging the
logs a bit more difficult, since you have to switch between the logs
printed in real times and the logs printed in the dump.

And in some cases, external users wish to redact their logs to the
minimum necessary and we've only received the resource update trace
dump. That means in those cases we're missing the direct log.

This change ensures the direct logs are in the dump buffer as well as
being written out immediately. They'll end up in the log twice if
there's a dump, but that's a reasonable tradeoff.

Bug: n/a
Test: n/a
Change-Id: I2dd890f6251652dadf251a9a3781fd0fa850269c

GitOrigin-RevId: ceb435a2e6c77d25094069de7345e2adea731857