Remove TaskGridItemModel and refactor + unify icon management
The TaskGridItemModel contains two fields: the ProfilerTaskType
enum and a path to the respective task icon. The addition of
the swing icon needed to be made to such grid item to be used
for the task tab's icon. However, that would mean defining
swing icons in the profilers module which is not allowed due to
the lack of the artwork module dependency. Thus, it would need
to be defined outside of the TaskGridItemModel and in the
profilers-ui module (causing redundant and fragmented icon
management). This naturally made it clear that the only data
needed to represent an item in the TaskGrid is the task type,
and all icons can be mapped to in the profilers-ui side of the
code. Icon paths were already managed by the TaskBasedUXIcons
in the profilers-ui module and the module itself is the only
profiler module that depends on the artwork module which contains
such icons. Thus, it made sense to use the TaskBasedUXIcons class
as the unified source of the swing icons, icon paths, and icon
classes.

Bug: 277798754
Test: existing
Change-Id: I690e4edf4ec9cbd5f39507588dae7be74c2089d4

GitOrigin-RevId: 9802cda2187664c6a75ffd68131cbecaf12be778