[android] Refine Kotlin plugin dependencies in intellij.android.projectSystem.gradle module
The `kotlin-gradle-plugin-idea` library, which is bundled with the Kotlin plugin, is currently used in the `intellij.android.projectSystem.gradle` module. However, the inclusion of this library led to test failures in `IdeaUltimatePluginStructureTest.plugins`, due to an `AssertionError` being raised. The error specifies that the 'kotlin-tooling-core' library is being included unnecessarily, as it's only used by a single plugin, and hence, suggests either converting it to a module-level library or using 'withProjectLibrary("kotlin-tooling-core")', to reduce the size of the platform distribution and prevent potential conflicts with bundled libraries in other plugins.

To address this issue, we have modified the dependency declaration for `kotlin-gradle-plugin-idea` in the Android module to have the `PROVIDED` scope, which means IDEA will not bundle it in the final distribution. We have also removed the unnecessary `kotlin-tooling-core` dependency.

(cherry picked from commit 469c30f138d02d34bac4bfee28c5e98c55ece6aa)

IJ-MR-120646

GitOrigin-RevId: a0b62f3a1dd7ea9587e0f1dcaf8652569f570968