Fix UI freeze caused by DeviceSpecInjectionContributor
The contributor calls slow methods to calculate the injection to be
provided, so it shouldn't be called from EDT, which happens in some
occasions (e.g. when creating a new project) when the Injection is not
needed.

To fix this issue, we return a null Injection when getInjection is
called from EDT.

Fixes: 312760215
Test: DeviceSpecInjectorTest
Change-Id: Ib5d30b9c369ea21128f3ef9fac051ca1f18fabb5

GitOrigin-RevId: 078057dad60402c63c4d36b1a5d409c3db5f9861