Ensure selection is preserved in Problems panel on refresh
When the problems tab for design tools is refreshed, the issue tree
tries to preserve selection of issues if one was selected before the
refresh. This relies on equality between issue nodes before and after
refresh.
For Visual Lint issues, this could fail if the issue in the node was
recreated. This changes the equals method for VisualLintIssueNode to
check whether the new node represents the same issue as the old one
instead of relying on object equality.

Bug: 312896127
Test: test added
Change-Id: I0aedb09ffed298b2bf4b9c435a62423fe37f3e30

GitOrigin-RevId: 1f158ba278ec61201bb5a4c94b67fb360293c367