Stop adding stdlib dependency on KotlinAndroidGradleModuleConfigurator
Starting with KGP 1.4.0 this dependency is under the hood by default,
meaning that not necessary to explicitly declare it during J2K
project configuration. In addition, this behaviour can be disabled,
if desired, by defining the following on gradle.properties.
- kotlin.stdlib.default.dependency=false

The configurator dialog displays the KGP versions that are obtained
from ConfigureDialogWithModulesAndVersion.VERSIONS_LIST_URL which
represents the latest 20 available versions from Maven. For this
reason is not expected to deal with any version lower than 1.4.0,
however, just in case a minimum supported version of 1.4.0 was defined
for KotlinAndroidGradleModuleConfigurator to skip any lower version.

Bug: 265906401
Test: ChangeProjectLanguageFeatureTest
Test: ConfigureProjectTest
Change-Id: I9569e521ba5b0025d4c66ff76aaccdcd365fa72b

GitOrigin-RevId: dc484a280cde03d0cfa33cba7288d75b7947eabe