Refactor LE tests to not re-create KtFiles
Updates the LE test "pattern" to involve creating and modifying KtFiles
in WriteActions, similar to how LE would be used in actuality.

This will prevent the PSI from getting upset that a single file is being
constantly recreated, and is a much cleaner pattern:

1. Create initial file with projectRule.createKtFile(name, content)
2. Modify with projectRule.modifyKtFile(file, content)
3. Creating an initial diff cache can be done with
   projectRule.initialCache(list of files)

This will additionally make it easier to write pseudo-E2E Live Edit
tests using the actual LE change detection + push implementation, as
well as allow for additional tests of the PSI validation.

Fixes: 303114659
Test: All tests pass
Change-Id: I003d60b49f8c27f94601572ae7b97cfd2317838d

GitOrigin-RevId: f95a07fc7682d4675559215ed3769a981b9ca1a3