Fix indefinite loading when there are no Compose Previews
In change If33c29edf, we fixed a zoom-to-fit action caused by an
inconsistency in the "onAfterRender" callback, which indirectly
called zoom-to-fit after the first render. The problem is that this
callback is also called if no previews were rendered, e.g. in the
initializing panel, where zoom-to-fit doesn't make any sense. By the
point it actually renders previews, we would think that the first
render had actually happened and would not zoom-to-fit.

The problem with the fix, which prevented onAfterRender from being
called if no previews were rendered, was that we would never leave
the "Initializing..." loading state for files with no previews. This
change moves the "previewCount > 0" check to onAfterRender itself, and
now we're only calling zoom-to-fit and updating hasRenderedAtLeastOnce
if rendering at least one preview. This will make sure that we won't
regress on bug 311056411, as well as will fix 312872445, since the
workbench's hasRendered (not to be confused with the preview state
mentioned above) state will be updated even if the file doesn't have
any previews.

Fixes: 312872445
Test: ComposePreviewRepresentationTest
Change-Id: Ic77e1c5aada5f9f3f1f19c8d6d5d7b3ac46cd60e

GitOrigin-RevId: 5c7e4a722869ad8e6e140e45adb8cebf1182c8a0