Change IdeChannel API to include target stability
By using isAtMost() and isAtLeast() instead of isLessStableThan() and
isMoreStableThan(), respectively, we can pass the intended channel as an
argument, which results in a better API usage. For example, if a feature
is intended to be enabled in canary and lower versions, we would simply
call isAtMost(CANARY), not needing to know the next stability version to
set isLessStableThan(BETA) instead.

Bug: N/A
Test: IdeChannelTest
Change-Id: If7a69a072305f0aa4fc45a3d4436dc4532545148

GitOrigin-RevId: 60c301850101697f9679aed5bffe95d04df4f669