Fix zoom to fit on entering default mode in Compose preview
For the zoom to fit to be correctly apply, any changes in preview
layouts have to happen before it is called in onEnter. This fixes two
issues:
- The change of layout must happen before onEnter, so that zoom to fit
computes the scale based on the new preview layout.
- In onUiCheckPreviewStop, setting the uiCheckFilter to disabled has to
be done before any ui thread calls to ensure that it will happen before
zoom to fit is called.

Bug: N/A
Test: Checked manually
Change-Id: I1bfd62d603cf039de6ddfd6937d22fdadce2de17

GitOrigin-RevId: 429bc8818fb4ea423c0043ec86718403b6015b87