Centrally deal with laying out invisible scene views
Instead of having each layout manager deal with invisible scene views
when laying out or measuring, this puts the base class in charge of
setting the location of all invisible scene views, and then delegating
the measuring and laying out of the visible views to subclasses.
This removes the danger of one layout implementation not correctly dealing
with invisible views.

Bug: 313984121
Test: test added
Change-Id: Iea7a331dffa321a1af4b21ea91d71bcedefd79d7

GitOrigin-RevId: 827032d2814f2a2e9a088b6ec6a043745a3af6c0