Fix missing navigatable for Compose NlComponent
The navigatable for an NlComponent build for Compose comes from matching
the information from AccessibilityNodeInfo and ComposeViewInfo. However,
this information can differ; in particular, bounds from Accessibility can
be larger than bounds from ComposeViewInfo. In order to get the correct
navigatable, we always look for it inside the main ComposeViewInfo root
bounds.

Bug: 312841771
Test: Checked manually
Change-Id: Ibd73a419ed24900672c9dde521223f6ea4db5f2e

GitOrigin-RevId: 3b39b579de6972a250a51d08bf4807c917227bdd