[Vcs Integration] Bring up the existing diff view if possible
...and scroll to the desired location.

This change followed how IJ brings up the existing diff view,
basically, PreviewDiffVirtualFile is what we need as it implements
equals which is checked when searching for the existing editor window.

As a nice side-effct, we no longer pre-check revision content to show
nicer error messages, instead we simply inject our message in the code.

Bug: 314198776
Test: added
Change-Id: I23e90c2b49793c4c82e3e7d6d1511db02867b8f1

GitOrigin-RevId: f9b26cf73079b1cc788ea924a3f4e4ed19d15e70