Disable Essentials Mode
Currently, the typing latency savings are neglible and user adoption is
very small so it does not make sense to move forward with it as is.
Using b/314837968 to decide what we want to do with the mode and
existing code in the future, but for now disabling the feature.

Bug: 314837968
Tests: Existing
Change-Id: I4af8ef053103a1d25733333f25bccc44eeaf4514

GitOrigin-RevId: 96564bc5f34cd70c44160f5ba0229e575a1bf5f2