Remove OldAndroidPluginIssueChecker and related code
This issue checker is obsolete. It checks for an error from tooling
API when Gradle version <2.6, it is the only place in Gradle codebase
where such error would be thrown. IJ now does not allow to run with
Gradle <3.0 so this error can never happen. Other incompatibility
errors are treated by other issue checkers.

Bug: 296841520
Test: N/A
Change-Id: I5c4ee76233cb1db032df7420040ffa2eca83047b

GitOrigin-RevId: 521b2f7749686a552cb52d1506c08e411fa03e2a